requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
asyncpg>=0.29.0
python-dotenv>=1.0.0
jinja2==3.1.3
duckdb>=1.1.0
anthropic>=0.40.0
# Research runner dependencies
scipy>=1.13.0
matplotlib>=3.9.0
reportlab>=4.2.0
pyyaml>=6.0.2
